// src/api/back/models/MatchPageDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import {
  MatchDto,
  MatchDtoFromJSON,
  MatchDtoFromJSONTyped,
  MatchDtoToJSON,
} from "./";

/**
 *
 * @export
 * @interface MatchPageDto
 */
export interface MatchPageDto {
  /**
   *
   * @type {Array<MatchDto>}
   * @memberof MatchPageDto
   */
  data: Array<MatchDto>;
  /**
   *
   * @type {number}
   * @memberof MatchPageDto
   */
  page: number;
  /**
   *
   * @type {number}
   * @memberof MatchPageDto
   */
  pages: number;
  /**
   *
   * @type {number}
   * @memberof MatchPageDto
   */
  perPage: number;
}

export function MatchPageDtoFromJSON(json: any): MatchPageDto {
  return MatchPageDtoFromJSONTyped(json, false);
}

export function MatchPageDtoFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): MatchPageDto {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    data: (json["data"] as Array<any>).map(MatchDtoFromJSON),
    page: json["page"],
    pages: json["pages"],
    perPage: json["perPage"],
  };
}

export function MatchPageDtoToJSON(value?: MatchPageDto | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    data: (value.data as Array<any>).map(MatchDtoToJSON),
    page: value.page,
    pages: value.pages,
    perPage: value.perPage,
  };
}